reorganising the code a bit, we still have lots of test to implement so we need to make a bit clearer
